% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HS_LUKK_Space}
\alias{HS_LUKK_Space}
\title{PhysioSpace Based on Human Gene Expression Atlas}
\format{A matrix with genes as rows and different physiological dimensions in columns.
Gene Entrez IDs are stored in rownames and space-dimension names in colnames.}
\source{
\url{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-62/}
}
\usage{
data(HS_LUKK_Space)
}
\description{
HS_LUKK_Space spans a 369-dimension physiological space based on the Lukk et. al. paper "A global map of human gene expression",
nature biotechnology 2010. Data downloaded from ArrayExpress, accession number E-MTAB-62.
}
\keyword{datasets}
